Descontinuar antigo main.py
